2.30.21
+ Evaluation : test si les reines peuvent subir une attaque à la découverte

		Results of Zangdar-2.30.21-g++ vs Zangdar-2.30.17-g++ (10+0.1, 1t, 128MB, openings-10ply-100k.pgn): concurrency 24
		Elo: 4.92 +/- 3.67, nElo: 8.03 +/- 6.00
		LOS: 99.56 %, DrawRatio: 46.49 %, PairsRatio: 1.08
		Games: 12864, Wins: 2762, Losses: 2580, Draws: 7522, Points: 6523.0 (50.71 %)
		Ptnml(0-2): [208, 1444, 2990, 1538, 252], WL/DD Ratio: 0.32
		LLR: 2.95 (-2.94, 2.94) [0.00, 5.00]

+++ src/MovePicker.cpp
@@ -361,7 +361,7 @@ int MovePicker::get_best(const MoveList& ml)
     int best_index = 0;
 
     // Find highest scoring move
-    for (int i = 1; i < ml.count; i++)
+    for (size_t i = 1; i < ml.count; i++)
     {
         if (ml.mlmoves[i].value > ml.mlmoves[best_index].value)
             best_index = i;
@@ -434,7 +434,7 @@ bool MovePicker::is_legal(MOVE move)
         gen_legal = true;
     }
 
-    for (int n=0; n<mll.count; n++)
+    for (size_t n=0; n<mll.count; n++)
         if (mll.mlmoves[n].move == move)
             return true;
     return false;
